Move to AssetManager. Remove struct ImagePicker
