+++ Source/BottomView/StackView.swift
@@ -140,7 +140,7 @@ extension ImageStackView {
 
     for (index, view) in views.enumerate() {
       if index <= photos.count - 1 {
-        ImagePicker.resolveAsset(photos[index], size: CGSize(width: Dimensions.imageSize, height: Dimensions.imageSize)) { image in
+        AssetManager.resolveAsset(photos[index], size: CGSize(width: Dimensions.imageSize, height: Dimensions.imageSize)) { image in
           view.image = image
         }
         view.alpha = 1
